+++ types/src/RandomTableResult.d.ts
@@ -3,15 +3,18 @@
  */
 export default class RandomTableResult {
     /**
+     * @property {String} key Key for RandomTable.
      * @property {String} table Title of subtable.
      * @property {String} result Randomized result label.
      * @property {String} desc Extra result description.
      */
-    constructor({ table, result, desc }: {
+    constructor({ key, table, result, desc }: {
+        key?: string;
         table?: string;
         result?: string;
         desc?: string;
     });
+    key: string;
     table: string;
     result: string;
     desc: string;